backend/requirements.txt
# FastAPI core
fastapi==0.115.11
uvicorn==0.34.0
starlette==0.46.0

# Validation
pydantic[email]==2.10.6
pydantic_core==2.27.2
annotated-types==0.7.0

# Database
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10

# Auth
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
cryptography==44.0.2
cffi==1.17.1
pycparser==2.22
asn1crypto==1.5.1

# Cloud & Storage
google==3.0.0
google-generativeai>=0.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
minio==7.2.15
cloudinary==1.44.0

# ML & Data Processing
faiss-cpu==1.11.0
numpy==2.2.6
pandas==2.2.3

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
typing_extensions==4.12.2
click==8.1.8
yt-dlp>=2024.12.6
colorama==0.4.6
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
idna==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tzdata==2025.2
urllib3==2.4.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
